Rename the live_model_creator.
We've come full circle, and I've re-instated the application model as the thing which has the
instances. It's also the thing that knows how to resolve etc. Conceptually it could be nicer
if some of the logic was pulled back out (I'm unsure), however the internal data structures
we're using will need some refactoring first.