Make actions get passed previous actions
This allows the wait actions to work out when they started waiting
and thus control their own timeout.

It also means that we can teach UpdateVersionAction that if it tries
more than a couple of times, it should give up, as the app is
seemingly never going to come up.

Once we've done that, we can fail faster, i.e. whenever we get an error
reported from an action, and we can change the action running loop
to be infinite (or nearer infinite)