Stop the cmdb location from being hard coded.
We will now load a config file from /Users/t0m/.orc.yaml which indicates the
CMDB repos location and where to keep the local checkout.

Current defaults are used if this file is not supplied.

CLI option parser logic was refactored to _always_ use the factory, thus making any modules
which the CLI uses directly benefit from the factory. Also, the factory is always constructed now,
as this enables us to add options to the CLI for things like the config file location which are
always optional, but have a global scope (i.e. them being passed down into an individual action
would be no help at all - ergo giving control of all the options to the factory allows it to control
argument passing down..)