gallan: add a command to query cmdb vs. the live model and indicate whether resolutions are needed.
In order to make this command useful in scripts, it was necessary to silence certain output. "git gc" was the worst offender, this will only be printed in debug mode (e.g. set --debug or -D). I imagine the git gc output was only there cause nobody needed to get rid of it, not that it was useful.

The other case to silence was the model builder, an opt-in "quiet" mode, that's only true for my new command, seemed better here.